% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_urls}
\alias{base_urls}
\title{Base URLs used in oncoEnrichR for linking out to external databases}
\format{
A list of characters strings with base URLs
}
\usage{
base_urls
}
\description{
Base URLs used in oncoEnrichR for linking out to external databases
}
\keyword{datasets}
