% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_drug.R
\name{cancer_association_rank}
\alias{cancer_association_rank}
\title{Rank a gene list according to cancer relevance to a tumor type/site}
\usage{
cancer_association_rank(
  query = NULL,
  tumor_site = "Breast",
  query_id_type = "symbol",
  ignore_id_err = TRUE,
  include_gene_summary = FALSE,
  oeDB = NULL
)
}
\arguments{
\item{query}{character vector with gene/query identifiers (minimum 1, maximum 2,500)}

\item{tumor_site}{character indicating primary tumor site of interest}

\item{query_id_type}{character indicating source of query (one of
"uniprot_acc", "symbol","entrezgene", or "ensembl_gene", "ensembl_mrna",
"refseq_transcript_id", "ensembl_protein", "refseq_protein")}

\item{ignore_id_err}{logical indicating if analysis should
continue when uknown query identifiers are encountered}

\item{include_gene_summary}{logical indicating if gene summary (NCBI/UniProt) should be included
in output data tables}

\item{oeDB}{oncoEnrichR data repository object - as returned from \code{load_db()}}
}
\value{
A list with two data frames: query genes ranked according to cancer relevance
for the specified tumor site, and unranked query genes (no evidence found)
}
\description{
Function that ranks a list of human gene identifiers with respect to
strength of association to a particular tumor type/site. Underlying association
evidence for the cancer rank per site is pulled from the Open Targets Platform,
see more details \href{https://sigven.github.io/oncoEnrichR/articles/cancer_gene_rank.html}{here}.
}
