% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Colors for tissue types in the gene fitness plot section in oncoEnrichR}
\format{
A list with two color code lists, one character vector of
hex color codes for tissues (gene fitness plot) and one data frame
with binned hex color codes for subcellular compartments
}
\usage{
color_palette
}
\description{
Colors for tissue types in the gene fitness plot section in oncoEnrichR
}
\keyword{datasets}
