% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp_output_cols}
\alias{cp_output_cols}
\title{Format (columns) for clusterProfiler enrichment output}
\format{
A character vector of 23 variables
}
\usage{
cp_output_cols
}
\description{
Format (columns) for clusterProfiler enrichment output
}
\keyword{datasets}
