% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{generate_report_data}
\alias{generate_report_data}
\title{Function that interrogates and analyzes a list of human protein-coding genes for cancer relevance}
\usage{
generate_report_data(
  query,
  query_source = "symbol",
  p_title = "Project Title",
  p_owner = "Project Owner",
  report_fname = "oncoEnrichR_Report",
  background_fname = NULL,
  background_enrichment = NULL,
  background_enrichment_source = "symbol",
  background_enrichment_description = "All protein-coding genes",
  p_value_cutoff_enrichment = 0.05,
  q_value_cutoff_enrichment = 0.01,
  ppi_add_nodes = 50,
  ppi_score_threshold = 900,
  show_ppi = T,
  show_disease = T,
  show_enrichment = T,
  show_tcga_aberration = T,
  show_tcga_coexpression = T,
  show_subcell_comp = T,
  show_complex = T
)
}
\arguments{
\item{query}{character vector with gene identifiers}

\item{query_source}{character indicating source of query (one of 'uniprot_acc', 'symbol', 'entrezgene', or 'ensembl_gene_id')}

\item{p_title}{title of report}

\item{p_owner}{name of project owner}

\item{report_fname}{filename (prefix) for HTML report and JSON output}

\item{background_fname}{filename for simple text file with project background information, one line per background item}

\item{background_enrichment}{character vector with gene identifiers, used as reference/background for enrichment analysis}

\item{background_enrichment_source}{character indicating source of background ('uniprot_acc','symbol','entrezgene','ensembl_gene_id')}

\item{background_enrichment_description}{character indicating type of background (e.g. 'All lipid-binding proteins (n = 200)')}

\item{p_value_cutoff_enrichment}{cutoff p-value for enrichment analysis}

\item{q_value_cutoff_enrichment}{cutoff q-value for enrichment analysis}

\item{ppi_add_nodes}{number of nodes to add to query set when computing the protein-protein interaction network}

\item{ppi_score_threshold}{minimum score (total) for included protein-protein interactions}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction data (STRING)}

\item{show_disease}{logical indicating if report should contain disease associations (Open Targets Platform)}

\item{show_enrichment}{logical indicating if report should contain functional enrichment analysis (MSigDB, GO, KEGG, REACTOME etc.)}

\item{show_tcga_aberration}{logical indicating if report should contain TCGA aberration plots (amplifications/deletions)}

\item{show_tcga_coexpression}{logical indicating if report should contain TCGA co-expression data (RNAseq) of queryset with oncogenes/tumor suppressor genes}

\item{show_subcell_comp}{logical indicating if report should list subcellular compartment annotations from ComPPI}

\item{show_complex}{logical indicating if report should list proteins in known protein complexes}

\item{format}{file format of output (html/json)}
}
\description{
Function that interrogates and analyzes a list of human protein-coding genes for cancer relevance
}
