% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{get_string_ppi_network}
\alias{get_string_ppi_network}
\title{Annotate geneset with protein complexes they are involved with}
\usage{
get_string_ppi_network(qgenes, genedb = NULL, settings = NULL)
}
\arguments{
\item{qgenes}{character vector of human gene symbols}

\item{genedb}{data frame with gene annotations}

\item{settings}{oncoEnrichR report settings}
}
\value{
data frame with unique protein complexes and their associated members
}
\description{
Annotate geneset with protein complexes they are involved with
}
