% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{load_db}
\alias{load_db}
\title{Load oncoEnrichR annotation database}
\usage{
load_db(github = T, cache_dir = NULL)
}
\arguments{
\item{github}{If TRUE (default), load database from GitHub repository}

\item{cache_dir}{path to local cache for more efficient subsequent loads}
}
\description{
Load oncoEnrichR annotation database
}
