% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release_notes}
\alias{release_notes}
\title{Versions/attributes of oncoEnrichR annotation datasets}
\format{
A list of lists, eech item being a list of five variables
\itemize{
\item \emph{url} - URL for annotation/software resource
\item \emph{description} - Description of resource
\item \emph{version} - version
\item \emph{name} - Name of resource
\item \emph{resource_type} - Type of resource (db/software)
}
}
\usage{
release_notes
}
\description{
Versions/attributes of oncoEnrichR annotation datasets
}
\keyword{datasets}
