% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subcell_map}
\alias{subcell_map}
\title{Format for subcell map object}
\format{
A list with three elements, a data frame with subcellular map
identities (go terms, UniProt ids etc), a data frame with styled UI
elements, and an SVG image of the subcellular map
}
\usage{
subcell_map
}
\description{
Format for subcell map object
}
\keyword{datasets}
