% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_colors}
\alias{tissue_colors}
\title{Colors for tissue types in the gene fitness plot section in oncoEnrichR}
\format{
A character vector of 24 color codes
}
\usage{
tissue_colors
}
\description{
Colors for tissue types in the gene fitness plot section in oncoEnrichR
}
\keyword{datasets}
