% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_db}
\alias{validate_db}
\title{Function that validates the oncoEnrichR database object}
\usage{
validate_db(oe_db)
}
\arguments{
\item{oe_db}{list object with annotation data for oncoEnrichR}
}
\description{
Function that validates the oncoEnrichR database object
}
\keyword{internal}
