% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_db_df}
\alias{validate_db_df}
\title{Function that validates a particular db object (data.frame) in oncoEnrichR}
\usage{
validate_db_df(df, dbtype = "genedb")
}
\arguments{
\item{df}{data.frame with annotation data for oncoEnrichR}

\item{dbtype}{type of oncoEnrichR datasource}
}
\description{
Function that validates a particular db object (data.frame) in oncoEnrichR
}
\keyword{SM_tractability_category',}
\keyword{gencode_gene_biotype',}
\keyword{internal}
\keyword{name',}
