% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{generate_report}
\alias{generate_report}
\title{Function that combines R markdown templates with the report object to produce and write an HTML report to file}
\usage{
generate_report(
  query,
  query_source = "symbol",
  p_title = "Project Title",
  p_owner = "Project Owner",
  report_fname = "OncoEnrichR_TestReport",
  background_fname = "moncho_background.txt",
  ppi_add_nodes = 50,
  background_enrichment = NULL,
  background_enrichment_source = "symbol",
  background_enrichment_description = "All protein-coding genes",
  show_ppi = T,
  show_disease = T,
  show_enrichment = T,
  show_tcga_aberration = T,
  show_tcga_coexp = T,
  show_gtex_coexp = F,
  show_complex = T,
  gtex_atlasassay_groups = c("g32", "g9", "g29", "g10", "g28", "g44", "g33", "g50",
    "g37", "g38", "g42", "g35")
)
}
\arguments{
\item{query}{character vector with gene identifiers}

\item{query_source}{character indicating source of query ('uniprot_acc','symbol','entrezgene','ensembl_gene_id')}

\item{p_title}{title of report}

\item{p_owner}{name of project owner}

\item{report_fname}{filename for report (no extension)}

\item{background_fname}{filename for file with project background information}

\item{ppi_add_nodes}{number of nodes to add to protein-protein interaction network}

\item{background_enrichment}{character vector with gene identifiers, used as reference/background for enrichment analysis}

\item{background_enrichment_source}{character indicating source of background ('uniprot_acc','symbol','entrezgene','ensembl_gene_id')}

\item{background_enrichment_description}{character indicating type of background (e.g. 'All lipid-binding proteins (n = 200)')}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction network}

\item{show_disease}{logical indicating if report should contain disease associations}

\item{show_enrichment}{logical indicating if report should contain enrichment analysis}

\item{show_tcga_aberration}{logical indicating if report should contain TCGA aberration plots}

\item{show_tcga_coexp}{logical indicating if report should contain co-expression data of queryset with oncogenes/tumor suppressor genes (TCGA)}

\item{show_gtex_coexp}{logical indicating if report should contain co-expression from GTex}

\item{show_complex}{logical indicating if report should list proteins in known protein complexes}

\item{gtex_atlasassay_groups}{character vector of GTex tissue identifiers (for co-expression plots)}

\item{report_name}{filename for report}

\item{format}{file format of output (html/json)}
}
\description{
Function that combines R markdown templates with the report object to produce and write an HTML report to file
}
