% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{init_report}
\alias{init_report}
\title{Function that initiates oncoEnrichR report structure}
\usage{
init_report(
  oeDB,
  project_title = "Project title",
  project_owner = "Project owner",
  html_floating_toc = T,
  html_report_theme = "default",
  query_id_type = "symbol",
  ignore_id_err = TRUE,
  project_description = "Project description",
  ppi_min_string_score = 900,
  ppi_add_nodes = 50,
  bgset_description = "All annotated protein-coding genes",
  bgset_id_type = "symbol",
  p_value_cutoff_enrichment = 0.05,
  p_value_adjustment_method = "BH",
  q_value_cutoff_enrichment = 0.2,
  min_geneset_size = 10,
  max_geneset_size = 500,
  num_terms_enrichment_plot = 20,
  min_subcellcomp_confidence = 1,
  subcellcomp_show_cytosol = F,
  min_confidence_reg_interaction = "D",
  simplify_go = F,
  show_ppi = T,
  show_drugs_in_ppi = T,
  show_disease = T,
  show_top_diseases_only = T,
  show_cancer_hallmarks = T,
  show_drug = T,
  show_enrichment = T,
  show_tcga_aberration = T,
  show_tcga_coexpression = T,
  show_subcell_comp = T,
  show_fitness = T,
  show_cell_tissue = F,
  show_ligand_receptor = T,
  show_regulatory_interactions = T,
  show_unknown_function = T,
  show_prognostic_cancer_assoc = T,
  show_complex = T
)
}
\arguments{
\item{oeDB}{oncoEnrichR annotation database object}

\item{project_title}{project title (title of report)}

\item{project_owner}{name of project owner}

\item{html_floating_toc}{logical - float the table of contents to the left of the main document content. The floating table of contents will always be visible even when the document is scrolled}

\item{html_report_theme}{Bootswatch theme for HTML report (any of "bootstrap","cerulean","cosmo","default","flatly","journal","lumen","paper","sandstone","simplex","spacelab","united","yeti")}

\item{query_id_type}{character indicating source of query (one of "uniprot_acc", "symbol",
"entrezgene", or "ensembl_gene","ensembl_mrna","refseq_mrna","ensembl_protein","refseq_protein")}

\item{ignore_id_err}{logical indicating if analysis should continue when uknown query identifiers are encountered}

\item{project_description}{project background information}

\item{ppi_min_string_score}{minimum score (0-1000) for retrieval of protein-protein interactions (STRING)}

\item{ppi_add_nodes}{number of nodes to add to target set when computing the protein-protein interaction network (STRING)}

\item{bgset_description}{character indicating type of background (e.g. "All lipid-binding proteins (n = 200)")}

\item{bgset_id_type}{character indicating source of query (one of "uniprot_acc", "symbol",
"entrezgene", or "ensembl_gene","ensembl_mrna","refseq_mrna","ensembl_protein","refseq_protein")}

\item{p_value_cutoff_enrichment}{cutoff p-value for enrichment/over-representation analysis}

\item{p_value_adjustment_method}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{q_value_cutoff_enrichment}{cutoff q-value for enrichment analysis}

\item{min_geneset_size}{minimal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{max_geneset_size}{maximal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{num_terms_enrichment_plot}{number of top enriched Gene Ontology terms (max) to show in enrichment barplot}

\item{min_subcellcomp_confidence}{minimum confidence level of subcellular compartment annotations (range from 1 to 6, 6 is strongest)}

\item{subcellcomp_show_cytosol}{logical indicating if subcellular heatmap should highlight cytosol as a subcellular protein location or not}

\item{min_confidence_reg_interaction}{minimum confidence level for regulatory interactions (TF-target) retrieved from DoRothEA ('A','B','C', or 'D')}

\item{simplify_go}{remove highly similar GO terms in results from GO enrichment/over-representation analysis}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction data (STRING)}

\item{show_drugs_in_ppi}{logical indicating if targeted drugs (> phase 3) should be displayed in protein-protein interaction network (Open Targets Platform)}

\item{show_disease}{logical indicating if report should contain disease associations (Open Targets Platform, association_score >= 0.05, support from at least two data types)}

\item{show_top_diseases_only}{logical indicating if report should contain top (n = 20) disease associations only pr. query gene (Open Targets Platform)}

\item{show_cancer_hallmarks}{logical indicating if report should contain annotations/evidence of cancer hallmarks per query gene (COSMIC/Open Targets Platform)}

\item{show_drug}{logical indicating if report should contain targeted cancer drug information}

\item{show_enrichment}{logical indicating if report should contain functional enrichment/over-representation analysis (MSigDB, GO, KEGG, REACTOME, NetPath, WikiPathways)}

\item{show_tcga_aberration}{logical indicating if report should contain TCGA aberration plots (amplifications/deletions)}

\item{show_tcga_coexpression}{logical indicating if report should contain TCGA co-expression data (RNAseq) of query set with oncogenes/tumor suppressor genes}

\item{show_subcell_comp}{logical indicating if report should provide subcellular compartment annotations (ComPPI)}

\item{show_fitness}{logical indicating if report should provide fitness scores and target priority scores from CRISPR/Cas9 loss-of-fitness screens (Project Score)}

\item{show_cell_tissue}{logical indicating if report should contain tissue-specificity and single cell-type specificity assessments (Human Protein Atlas)
of target genes, using data from the Human Protein Atlas}

\item{show_ligand_receptor}{logical indicating if report should contain ligand-receptor interactions (CellChatDB)}

\item{show_regulatory_interactions}{logical indicating if report should contain data on transcription factor (TF) - target interactions relevant for the query set (DoRothEA)}

\item{show_unknown_function}{logical indicating if report should highlight target genes with unknown or poorly defined functions (GO/Uniprot KB/NCBI)}

\item{show_prognostic_cancer_assoc}{logical indicating if mRNA-based (single-gene) prognostic associations to cancer types should be listed (Human Protein Atlas/TCGA)}

\item{show_complex}{logical indicating if report should provide target memberships in known protein complexes (ComplexPortal/Compleat/PDB/CORUM)}

\item{show_syn_leth_pairs}{logical indicating if report should list overlap with known synthetic lethality gene pairs (SynLethDB)}
}
\description{
Function that initiates oncoEnrichR report structure
}
\keyword{internal}
