% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{load_db}
\alias{load_db}
\title{Load oncoEnrichR datasets}
\usage{
load_db(cache_dir = NA, force_download = F)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should force downloaded
(i.e. set to TRUE to re-download even if data exists in cache)}
}
\description{
Function that loads version-tagged oncoEnrichR data from Google Drive
}
