% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{onco_enrich}
\alias{onco_enrich}
\title{Interrogate a gene list for cancer relevance}
\usage{
onco_enrich(
  query = NULL,
  oeDB = NULL,
  query_id_type = "symbol",
  ignore_id_err = TRUE,
  html_floating_toc = T,
  html_report_theme = "default",
  project_title = "_Project title_",
  project_owner = "_Project owner_",
  project_description = "_Project description_",
  bgset = NULL,
  bgset_id_type = "symbol",
  bgset_description = "All protein-coding genes",
  p_value_cutoff_enrichment = 0.05,
  p_value_adjustment_method = "BH",
  q_value_cutoff_enrichment = 0.2,
  min_geneset_size = 10,
  max_geneset_size = 500,
  num_terms_enrichment_plot = 20,
  min_subcellcomp_confidence = 1,
  subcellcomp_show_cytosol = FALSE,
  min_confidence_reg_interaction = "D",
  max_fitness_score = -2,
  simplify_go = TRUE,
  ppi_add_nodes = 50,
  ppi_score_threshold = 900,
  ppi_node_shadow = TRUE,
  show_ppi = TRUE,
  show_drugs_in_ppi = TRUE,
  show_disease = TRUE,
  show_top_diseases_only = TRUE,
  show_cancer_hallmarks = TRUE,
  show_drug = TRUE,
  show_enrichment = TRUE,
  show_aberration = TRUE,
  show_coexpression = TRUE,
  show_cell_tissue = FALSE,
  show_ligand_receptor = TRUE,
  show_regulatory = TRUE,
  show_unknown_function = TRUE,
  show_prognostic = TRUE,
  show_subcell_comp = TRUE,
  show_synleth = TRUE,
  show_fitness = TRUE,
  show_complex = TRUE,
  show_domain = TRUE,
  ...
)
}
\arguments{
\item{query}{character vector with gene/query identifiers}

\item{oeDB}{oncoEnrichR data repository object - as returned from \code{load_db()}}

\item{query_id_type}{character indicating source of query (one of "uniprot_acc", "symbol",
"entrezgene", or "ensembl_gene", "ensembl_mrna", "refseq_mrna", "ensembl_protein", "refseq_protein")}

\item{ignore_id_err}{logical indicating if analysis should continue when uknown query identifiers are encountered}

\item{html_floating_toc}{logical - float the table of contents to the left of the main document content (HTML report). The floating table of contents will always be visible even when the document is scrolled}

\item{html_report_theme}{Bootswatch theme for HTML report (any of "bootstrap", "cerulean", "cosmo", "default",
"flatly", "journal", "lumen", "paper", "sandstone", "simplex", "spacelab", "united", "yeti")}

\item{project_title}{project title (title of report)}

\item{project_owner}{name of project owner}

\item{project_description}{project background information}

\item{bgset}{character vector with gene identifiers, used as reference/background for enrichment/over-representation analysis}

\item{bgset_id_type}{character indicating source of background ("uniprot_acc", "symbol", "entrezgene",
"ensembl_gene", "ensembl_mrna", "refseq_mrna", "ensembl_protein", "refseq_protein")}

\item{bgset_description}{character indicating type of background (e.g. "All lipid-binding proteins (n = 200)")}

\item{p_value_cutoff_enrichment}{cutoff p-value for enrichment/over-representation analysis}

\item{p_value_adjustment_method}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{q_value_cutoff_enrichment}{cutoff q-value for enrichment analysis}

\item{min_geneset_size}{minimal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{max_geneset_size}{maximal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{num_terms_enrichment_plot}{number of top enriched Gene Ontology terms (max) to show in enrichment barplot}

\item{min_subcellcomp_confidence}{minimum confidence level of subcellular compartment annotations (range from 1 to 6, 6 is strongest)}

\item{subcellcomp_show_cytosol}{logical indicating if subcellular heatmap should show highlight proteins located in the cytosol or not}

\item{min_confidence_reg_interaction}{minimum confidence level for regulatory interactions (TF-target) retrieved from DoRothEA ('A','B','C', or 'D')}

\item{max_fitness_score}{maximum loss-of-fitness score (scaled Bayes factor from BAGEL) for genes retrieved from Project Score}

\item{simplify_go}{remove highly similar GO terms in results from GO enrichment/over-representation analysis}

\item{ppi_add_nodes}{number of nodes to add to target set when computing the protein-protein interaction network (STRING)}

\item{ppi_score_threshold}{minimum score (0-1000) for retrieval of protein-protein interactions (STRING)}

\item{ppi_node_shadow}{show shadow for nodes in the displayed PPI network}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction data (STRING)}

\item{show_drugs_in_ppi}{logical indicating if targeted drugs (> phase 3) should be displayed in protein-protein interaction network (Open Targets Platform)}

\item{show_disease}{logical indicating if report should contain disease associations (Open Targets Platform, association_score >= 0.05, support from at least two data types)}

\item{show_top_diseases_only}{logical indicating if report should contain top (n = 20) disease associations only pr. query gene (Open Targets Platform)}

\item{show_cancer_hallmarks}{logical indicating if report should contain annotations/evidence of cancer hallmarks per query gene (COSMIC/Open Targets Platform)}

\item{show_drug}{logical indicating if report should contain targeted cancer drug information}

\item{show_enrichment}{logical indicating if report should contain functional enrichment/over-representation analysis (MSigDB, GO, KEGG, REACTOME, NetPath, WikiPathways)}

\item{show_aberration}{logical indicating if report should contain TCGA aberration plots (amplifications/deletions)}

\item{show_coexpression}{logical indicating if report should contain TCGA co-expression data (RNAseq) of query set with oncogenes/tumor suppressor genes}

\item{show_cell_tissue}{logical indicating if report should contain tissue-specificity and single cell-type specificity assessments (Human Protein Atlas)
of target genes, using data from the Human Protein Atlas}

\item{show_ligand_receptor}{logical indicating if report should contain ligand-receptor interactions (CellChatDB)}

\item{show_regulatory}{logical indicating if report should contain data on transcription factor (TF) - target interactions relevant for the query set (DoRothEA)}

\item{show_unknown_function}{logical indicating if report should highlight target genes with unknown or poorly defined functions (GO/Uniprot KB/NCBI)}

\item{show_prognostic}{logical indicating if mRNA-based (single-gene) prognostic associations to cancer types should be listed (Human Protein Atlas/TCGA)}

\item{show_subcell_comp}{logical indicating if report should provide subcellular compartment annotations (ComPPI)}

\item{show_synleth}{logical indicating if report should list overlap with predicted synthetic lethality interactions (gene paralogs only, De Kegel et al., Cell Systems, 2021)}

\item{show_fitness}{logical indicating if report should provide fitness scores and target priority scores from CRISPR/Cas9 loss-of-fitness screens (Project Score)}

\item{show_complex}{logical indicating if report should provide target memberships in known protein complexes (ComplexPortal/Compleat/PDB/CORUM)}

\item{show_domain}{logical indicating if report should provide target memberships in known protein domains (Pfam)}

\item{...}{arguments for Galaxy/web-based processing}
}
\value{
An oncoEnrichR report list object, with two main elements,
\code{data} and \code{config}. \code{data} contains data that goes into
each section of the output reports, \code{config} contains all
metadata for annotation resources used.
}
\description{
Function that interrogates a list of human gene identifiers for
cancer relevance. Multiple perspectives are offered, including
tumor aberration and co-expression patterns, druggability,
protein-protein interactions, gene fitness effects, regulatory
interactions, subcellular compartment enrichment, pathway enrichment,
synthetic lethality interactions, prognostic associations, and more.
}
