% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{onco_enrich}
\alias{onco_enrich}
\title{Function that interrogates and analyzes a list of human protein-coding genes for cancer relevance}
\usage{
onco_enrich(
  query,
  query_source = "symbol",
  ignore_unknown = FALSE,
  p_title = "Project Title",
  p_owner = "Project Owner",
  background_fname = NULL,
  background_enrichment = NULL,
  background_enrichment_source = "symbol",
  background_enrichment_description = "All protein-coding genes",
  p_value_cutoff_enrichment = 0.05,
  p_value_adjustment_method = "BH",
  q_value_cutoff_enrichment = 0.2,
  min_geneset_size = 10,
  max_geneset_size = 500,
  simplify_go = F,
  ppi_add_nodes = 50,
  ppi_score_threshold = 900,
  show_ppi = T,
  show_drugs_in_ppi = F,
  show_disease = T,
  show_enrichment = T,
  show_tcga_aberration = T,
  show_tcga_coexpression = T,
  show_subcell_comp = T,
  show_crispr_lof = T,
  show_complex = T
)
}
\arguments{
\item{query}{character vector with gene/query identifiers}

\item{query_source}{character indicating source of query (one of 'uniprot_acc', 'symbol', 'entrezgene', or 'ensembl_gene_id')}

\item{ignore_unknown}{logical indicating if analysis should continue when uknown query identifiers are encountered}

\item{p_title}{title of report}

\item{p_owner}{name of project owner}

\item{background_fname}{filename for simple text file with project background information, one line per background item}

\item{background_enrichment}{character vector with gene identifiers, used as reference/background for enrichment/over-representation analysis}

\item{background_enrichment_source}{character indicating source of background ('uniprot_acc','symbol','entrezgene','ensembl_gene_id')}

\item{background_enrichment_description}{character indicating type of background (e.g. 'All lipid-binding proteins (n = 200)')}

\item{p_value_cutoff_enrichment}{cutoff p-value for enrichment/over-representation analysis}

\item{p_value_adjustment_method}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{q_value_cutoff_enrichment}{cutoff q-value for enrichment analysis}

\item{min_geneset_size}{minimal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{max_geneset_size}{maximal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{simplify_go}{remove highly similar GO terms in results from GO enrichment/over-representation analysis}

\item{ppi_add_nodes}{number of nodes to add to query set when computing the protein-protein interaction network}

\item{ppi_score_threshold}{minimum score (total) for included protein-protein interactions}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction data (STRING)}

\item{show_drugs_in_ppi}{logical indicating if targeted drugs (> phase 3) should be displayed in protein-protein interaction network}

\item{show_disease}{logical indicating if report should contain disease associations (Open Targets Platform)}

\item{show_enrichment}{logical indicating if report should contain functional enrichment/over-representation analysis (MSigDB, GO, KEGG, REACTOME etc.)}

\item{show_tcga_aberration}{logical indicating if report should contain TCGA aberration plots (amplifications/deletions)}

\item{show_tcga_coexpression}{logical indicating if report should contain TCGA co-expression data (RNAseq) of queryset with oncogenes/tumor suppressor genes}

\item{show_subcell_comp}{logical indicating if report should list subcellular compartment annotations from ComPPI}

\item{show_crispr_lof}{logical indicating if report should list results from CRISPR/Cas9 loss-of-fitness screens}

\item{show_complex}{logical indicating if report should list proteins in known protein complexes}
}
\description{
Function that interrogates and analyzes a list of human protein-coding genes for cancer relevance
}
