% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{onco_enrich}
\alias{onco_enrich}
\title{Interrogate a gene list for cancer relevance}
\usage{
onco_enrich(
  query,
  query_id_type = "symbol",
  ignore_id_err = TRUE,
  html_floating_toc = T,
  html_report_theme = "default",
  project_title = "Project title",
  project_owner = "Project owner",
  project_description = "Project description",
  bgset = NULL,
  bgset_id_type = "symbol",
  bgset_description = "All protein-coding genes",
  p_value_cutoff_enrichment = 0.05,
  p_value_adjustment_method = "BH",
  q_value_cutoff_enrichment = 0.2,
  min_geneset_size = 10,
  max_geneset_size = 500,
  min_subcellcomp_confidence = 1,
  subcellcomp_show_cytosol = FALSE,
  min_confidence_reg_interaction = "D",
  simplify_go = TRUE,
  ppi_add_nodes = 50,
  ppi_score_threshold = 900,
  show_ppi = TRUE,
  show_drugs_in_ppi = TRUE,
  show_disease = TRUE,
  show_top_diseases_only = TRUE,
  show_cancer_hallmarks = TRUE,
  show_drug = TRUE,
  show_enrichment = TRUE,
  show_tcga_aberration = TRUE,
  show_tcga_coexpression = TRUE,
  show_cell_tissue = FALSE,
  show_ligand_receptor = TRUE,
  show_regulatory_interactions = TRUE,
  show_unknown_function = TRUE,
  show_prognostic_cancer_assoc = TRUE,
  show_subcell_comp = TRUE,
  show_crispr_lof = TRUE,
  show_complex = TRUE,
  ...
)
}
\arguments{
\item{query}{character vector with gene/query identifiers}

\item{query_id_type}{character indicating source of query (one of "uniprot_acc", "symbol",
"entrezgene", or "ensembl_gene","ensembl_mrna","refseq_mrna","ensembl_protein","refseq_protein")}

\item{ignore_id_err}{logical indicating if analysis should continue when uknown query identifiers are encountered}

\item{html_floating_toc}{logical - float the table of contents to the left of the main document content (HTML report). The floating table of contents will always be visible even when the document is scrolled}

\item{html_report_theme}{Bootswatch theme for HTML report (any of "bootstrap","cerulean","cosmo","default","flatly","journal","lumen","paper","sandstone","simplex","spacelab","united","yeti")}

\item{project_title}{project title (title of report)}

\item{project_owner}{name of project owner}

\item{project_description}{project background information}

\item{bgset}{character vector with gene identifiers, used as reference/background for enrichment/over-representation analysis}

\item{bgset_id_type}{character indicating source of background ("uniprot_acc","symbol","entrezgene","ensembl_gene_id")}

\item{bgset_description}{character indicating type of background (e.g. "All lipid-binding proteins (n = 200)")}

\item{p_value_cutoff_enrichment}{cutoff p-value for enrichment/over-representation analysis}

\item{p_value_adjustment_method}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{q_value_cutoff_enrichment}{cutoff q-value for enrichment analysis}

\item{min_geneset_size}{minimal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{max_geneset_size}{maximal size of geneset annotated by term for testing in enrichment/over-representation analysis}

\item{min_subcellcomp_confidence}{minimum confidence level of subcellular compartment annotations (range from 1 to 6, 6 is strongest)}

\item{subcellcomp_show_cytosol}{logical indicating if subcellular heatmap should show highlight proteins located in the cytosol or not}

\item{min_confidence_reg_interaction}{minimum confidence level for regulatory interactions (TF-target) retrieved from DoRothEA ('A','B','C', or 'D')}

\item{simplify_go}{remove highly similar GO terms in results from GO enrichment/over-representation analysis}

\item{ppi_add_nodes}{number of nodes to add to target set when computing the protein-protein interaction network (STRING)}

\item{ppi_score_threshold}{minimum score (0-1000) for retrieval of protein-protein interactions (STRING)}

\item{show_ppi}{logical indicating if report should contain protein-protein interaction data (STRING)}

\item{show_drugs_in_ppi}{logical indicating if targeted drugs (> phase 3) should be displayed in protein-protein interaction network (Open Targets Platform)}

\item{show_disease}{logical indicating if report should contain disease associations (Open Targets Platform, association_score >= 0.4)}

\item{show_top_diseases_only}{logical indicating if report should contain top (15) disease associations only (Open Targets Platform)}

\item{show_cancer_hallmarks}{logical indicating if report should contain annotations/evidence of cancer hallmarks per query gene (COSMIC/Open Targets Platform)}

\item{show_enrichment}{logical indicating if report should contain functional enrichment/over-representation analysis (MSigDB, GO, KEGG, REACTOME, NetPath, WikiPathways)}

\item{show_tcga_aberration}{logical indicating if report should contain TCGA aberration plots (amplifications/deletions)}

\item{show_tcga_coexpression}{logical indicating if report should contain TCGA co-expression data (RNAseq) of query set with oncogenes/tumor suppressor genes}

\item{show_cell_tissue}{logical indicating if report should contain tissue-specificity and single cell-type specificity assessments (Human Protein Atlas)
of target genes, using data from the Human Protein Atlas}

\item{show_ligand_receptor}{logical indicating if report should contain ligand-receptor interactions (CellChatDB)}

\item{show_regulatory_interactions}{logical indicating if report should contain data on transcription factor (TF) - target interactions relevant for the query set (DoRothEA)}

\item{show_unknown_function}{logical indicating if report should highlight target genes with unknown or poorly defined functions (GO/Uniprot KB/NCBI)}

\item{show_prognostic_cancer_assoc}{logical indicating if mRNA-based (single-gene) prognostic associations to cancer types should be listed (Human Protein Atlas/TCGA)}

\item{show_subcell_comp}{logical indicating if report should provide subcellular compartment annotations (ComPPI)}

\item{show_crispr_lof}{logical indicating if report should provide fitness scores and target priority scores from CRISPR/Cas9 loss-of-fitness screens (Project Score)}

\item{show_complex}{logical indicating if report should provide target memberships in known protein complexes (ComplexPortal/Compleat/PDB/CORUM)}
}
\description{
Interrogate a gene list for cancer relevance
}
