% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{write}
\alias{write}
\title{Write oncoEnrichR report object to output file}
\usage{
write(
  report,
  oeDB,
  file = "testReport.html",
  ignore_file_extension = F,
  overwrite = F,
  format = "html",
  ...
)
}
\arguments{
\item{report}{object with oncoEnrichR report data (returned by oeDB$onco_enrich)}

\item{oeDB}{oncoEnrichR data repository object - as returned from \code{load_db()}}

\item{file}{full filename for report output (e.g. "oe_report.html" or "oe_report.xlsx")}

\item{ignore_file_extension}{logical to accept any type of filaname extensions (for Galaxy integration)}

\item{overwrite}{logical indicating if existing output files may be overwritten}

\item{format}{file format of output (html/excel)}

\item{...}{options for Galaxy/non self-contained HTML. Only applicable for use in Galaxy}
}
\description{
Function that writes an oncoEnrichR report object to file,
either as an interactive HTML report or as an Excel workbook.
}
