% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onco_enrichr.R
\name{write}
\alias{write}
\title{Function that writes the contents in the oncoEnrichR report object to an interactive HTML report}
\usage{
write(report, project_directory, report_name, format = "html")
}
\arguments{
\item{report}{object with oncoEnrichR report data (returned by oncoEnrichR::onco_enrich)}

\item{project_directory}{working directory}

\item{report_name}{filename for report}

\item{format}{file format of output (html/json)}
}
\description{
Function that writes the contents in the oncoEnrichR report object to an interactive HTML report
}
